requirements.txt
# Flask and API
Flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9

# Supabase
supabase==2.0.0
PyJWT==2.8.0

# Vector DB
pinecone-client==3.0.0
pinecone

# Async Tasks
celery==5.3.4
redis==5.0.1

# Text Processing
pdfplumber==0.10.3
beautifulsoup4==4.12.2
lxml==4.9.3

# HTTP Requests
requests==2.31.0

# Security
cryptography==41.0.7

# S3 (optional for Supabase Storage via S3 protocol)
boto3==1.34.34

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# CLI
click==8.1.7

# Environment
python-dotenv==1.0.0

